Pull out retry and add some tests
